+++ android/src/main/java/com/getcapacitor/community/media/photoviewer/fragments/ScreenSlidePageFragment.kt
@@ -1,7 +1,9 @@
 package com.getcapacitor.community.media.photoviewer.fragments
 
 import android.content.Context
+import android.net.Uri
 import android.os.Bundle
+import android.os.Environment
 import android.util.Log
 import android.view.KeyEvent
 import android.view.LayoutInflater
@@ -24,6 +26,7 @@ import com.getcapacitor.community.media.photoviewer.helper.BackgroundColor
 import com.getcapacitor.community.media.photoviewer.helper.CallbackListener
 import com.getcapacitor.community.media.photoviewer.helper.GlideApp
 import com.getcapacitor.community.media.photoviewer.helper.ShareImage
+import java.io.File
 
 
 class ScreenSlidePageFragment() : Fragment(), CallbackListener {
@@ -103,11 +106,32 @@ class ScreenSlidePageFragment() : Fragment(), CallbackListener {
         ivFullscreenImage.setBackgroundResource(backColor)
         tvGalleryTitle.text = image.title
 
-        GlideApp.with(appContext)
+        val imgUrl = image.url
+        if (imgUrl?.substring(0, 4).equals("http")) {
+          // load image from http
+          GlideApp.with(appContext)
             .load(image.url)
             .fitCenter()
             .diskCacheStrategy(DiskCacheStrategy.ALL)
             .into(ivFullscreenImage)
+        }
+        if (imgUrl?.substring(0, 4).equals("file")) {
+          val uri: Uri = Uri.parse(imgUrl)
+          val element: String? = uri.getLastPathSegment()
+          var file: File? = null
+          if (imgUrl?.contains("DCIM") == true) {
+            file = File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM).toString(), element )
+          }
+          if (imgUrl?.contains("Pictures") == true) {
+            file = File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES).toString(), element )
+          }
+          GlideApp.with(appContext)
+            .asBitmap()
+            .load(file)
+            .fitCenter()
+            .diskCacheStrategy(DiskCacheStrategy.ALL)
+            .into(ivFullscreenImage)
+        }
 
         val share: ImageButton = binding.shareBtn
         val close: ImageButton = binding.closeBtn