+++ android/src/main/java/com/getcapacitor/community/media/photoviewer/fragments/TouchViewFragment.kt
@@ -1,7 +1,9 @@
 package com.getcapacitor.community.media.photoviewer.fragments
 
 import android.content.Context
+import android.net.Uri
 import android.os.Bundle
+import android.os.Environment
 import android.util.Log
 import android.view.*
 import android.widget.RelativeLayout
@@ -14,6 +16,7 @@ import com.getcapacitor.community.media.photoviewer.helper.BackgroundColor
 import com.getcapacitor.community.media.photoviewer.helper.CallbackListener
 import com.getcapacitor.community.media.photoviewer.helper.GlideApp
 import com.ortiz.touchview.TouchImageView
+import java.io.File
 
 
 class TouchViewFragment(private val callbackListener: CallbackListener): DialogFragment() {
@@ -49,11 +52,32 @@ class TouchViewFragment(private val callbackListener: CallbackListener): DialogF
         ivTouchImage = binding.ivTouchImage
         appContext = this.requireContext()
 
-        GlideApp.with(appContext)
+        if (url.substring(0, 4).equals("http")) {
+          // load image from http
+          GlideApp.with(appContext)
             .load(url)
             .fitCenter()
             .diskCacheStrategy(DiskCacheStrategy.ALL)
             .into(ivTouchImage)
+        }
+        if (url.substring(0, 4).equals("file")) {
+          val uri: Uri = Uri.parse(url)
+          val element: String? = uri.getLastPathSegment()
+          var file: File? = null
+          if (url.contains("DCIM") == true) {
+            file = File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM).toString(), element )
+          }
+          if (url.contains("Pictures") == true) {
+            file = File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES).toString(), element )
+          }
+          GlideApp.with(appContext)
+            .asBitmap()
+            .load(file)
+            .fitCenter()
+            .diskCacheStrategy(DiskCacheStrategy.ALL)
+            .into(ivTouchImage)
+        }
+
         container?.addView(binding.root)
 
         return binding.root