+++ ios/Plugin/PhotoSlider/ImageScrollViewController.swift
@@ -136,7 +136,19 @@ class ImageScrollViewController: UIViewController {
 
         view.addSubview(mScrollView)
         mScrollView.addSubview(mImageView)
-        mImageView.sd_setImage(with: URL(string: url), placeholderImage: nil)
+        if url.prefix(4) == "http" {
+            mImageView.sd_setImage(with: URL(string: url),
+                                  placeholderImage: nil)
+        }
+        if url.prefix(38) ==
+                "file:///var/mobile/Media/DCIM/100APPLE" {
+            
+            mImageView
+                .getImageFromInternalUrl(url: url,
+                                         imgPlaceHolder: nil)
+        }
+
+
         mImageView.translatesAutoresizingMaskIntoConstraints = false
         addGestureRecognizers()
     }